backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "ai-english-mentor-backend"
version = "0.1.0"
description = "AI English Mentor backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "argon2-cffi>=23.1.0",
  "asyncpg>=0.30.0",
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "passlib[argon2]>=1.7.4",
  "pydantic-settings>=2.8.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.37",
  "uvicorn[standard]>=0.34.0",
  "requests>=2.32.3"
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.20.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.2",
  "pytest-cov>=6.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["app"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]
